+++ unit_tests/platforms/platform_sh/test_platformsh_config.py
@@ -88,6 +88,16 @@ def test_log_dir(tmp_project):
     assert "INFO: Logging run of `manage.py simple_deploy`..." in log_file_text
     assert "INFO: Configuring project for deployment to Platform.sh..." in log_file_text
 
+    assert "INFO: CLI args: {" in log_file_text
+    assert "INFO:   Deployment target: platform_sh" in log_file_text
+    assert "INFO:   Project name: blog" in log_file_text
+    assert "INFO: git status:" in log_file_text
+    assert "INFO: Untracked files:" in log_file_text
+    assert 'INFO:   (use "git add <file>..." to include in what will be committed)' in log_file_text
+    assert "INFO: \tsimple_deploy_logs/" in log_file_text
+
     # Spot check for success messages.
     assert "INFO: --- Your project is now configured for deployment on Platform.sh. ---" in log_file_text
     assert "INFO: To deploy your project, you will need to:" in log_file_text
+
+    assert "INFO: - You can find a full record of this configuration in the simple_deploy_logs directory." in log_file_text